updated for-loop index type to size_t

+++ assn4/prob2/src/Image.h
@@ -55,11 +55,11 @@ public:
 	Image(size_t _width, size_t _height) : m_width(_width), m_height(_height), m_buff(new PixelType[_width * _height]) { }
 
 	Image(size_t _width, size_t _height, const PixelType &val) : m_width(_width), m_height(_height), m_buff(new PixelType[_width * _height]) {
-		for (int i = 0; i < m_width * m_height; i++) m_buff[i] = val;
+		for (size_t i = 0; i < m_width * m_height; i++) m_buff[i] = val;
 	}
 
 	Image(const Image<PixelType> &img) : m_width(img.width), m_height(img.height), m_buff(new PixelType[img.width * img.height]) {
-		for (int i = 0; i < m_width * m_height; i++) m_buff[i] = img.m_buff[i];
+		for (size_t i = 0; i < m_width * m_height; i++) m_buff[i] = img.m_buff[i];
 	}
 
 	~Image() { }
@@ -71,14 +71,14 @@ public:
 	Image& operator=(const Image &img) {
 		m_width = img.m_width, m_height = img.m_height;
 		m_buff = PixelArray(new PixelType[m_width * m_height]);
-		for (int i = 0; i < m_width * m_height; i++) m_buff[i] = img.m_buff[i];
+		for (size_t i = 0; i < m_width * m_height; i++) m_buff[i] = img.m_buff[i];
 		return *this;
 	}
 	
 	const Image& operator=(const Image &img) const {
 		m_width = img.m_width, m_height = img.m_height;
 		m_buff = PixelArray(new PixelType[m_width * m_height]);
-		for (int i = 0; i < m_width * m_height; i++) m_buff[i] = img.m_buff[i];
+		for (size_t i = 0; i < m_width * m_height; i++) m_buff[i] = img.m_buff[i];
 		return *this;
 	}
